#26 - Seeding Of All The Data Added; Some Corrections To Statistics Entity; Migrations Renewed

+++ Query/Statistics/GetChartData/ChartDataDto.cs
@@ -8,7 +8,7 @@ namespace Query.Statistics.GetChartData
 {
     public class ChartDataDto
     {
-        public int SymbolsPerMinute { get; set; }
+        public float SymbolsPerMinute { get; set; }
         public DateTimeOffset Date { get; set; }
     }
 }
